slight refactoring and create overview page.
