+++ sources/lusan/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(LUSAN "${LUSAN_BASE}/lusan")
 
-include(${LUSAN}/application/CMakeLists.txt)
+include(${LUSAN}/app/CMakeLists.txt)
 include(${LUSAN}/common/CMakeLists.txt)
 include(${LUSAN}/res/CMakeLists.txt)
+include(${LUSAN}/model/CMakeLists.txt)
+include(${LUSAN}/view/CMakeLists.txt)
