+++ sources/lusan/app/CMakeLists.txt
@@ -1,0 +1,9 @@
+list(APPEND LUSAN_SRC
+    ${LUSAN}/app/LusanApplication.cpp
+    ${LUSAN}/app/main.cpp
+)
+
+list(APPEND LUSAN_HDR
+    ${LUSAN}/app/LusanApplication.cpp
+    ${LUSAN}/app/main.cpp
+)
