--- sources/lusan/application/CMakeLists.txt
@@ -1,2 +1,0 @@
-include("${LUSAN}/application/main/CMakeLists.txt")
-include("${LUSAN}/application/si/CMakeLists.txt")
