--- sources/lusan/application/main/ui/CMakeLists.txt
@@ -1,4 +1,0 @@
-list(APPEND LUSAN_UI
-    ${LUSAN}/application/main/ui/mainwindow.ui
-    ${LUSAN}/application/main/ui/workspace.ui
-)
