--- sources/lusan/application/si/CMakeLists.txt
@@ -1,9 +1,0 @@
-list(APPEND LUSAN_UI
-    ${LUSAN}/application/si/ui/overview.qml
-    ${LUSAN}/application/si/ui/sioverview.ui
-)
-
-list(APPEND LUSAN_SRC
-    ${LUSAN}/application/si/ServiceInterfaceView.cpp
-    ${LUSAN}/application/si/ui/sioverview.cpp
-)
