--- sources/lusan/application/si/ui/overview.qml
@@ -1,14 +1,0 @@
-import QtQuick 2.15
-
-Item {
-    visible: false
-    
-    Row {
-        id: row
-        x: 0
-        y: 8
-        width: 640
-        height: 464
-    }
-    
-}
