--- sources/lusan/application/si/ui/sioverview.cpp
@@ -1,14 +1,0 @@
-#include "sioverview.hpp"
-#include "ui_sioverview.h"
-
-sioverview::sioverview(QWidget *parent)
-    : QWidget(parent)
-    , ui(new Ui::sioverview)
-{
-    ui->setupUi(this);
-}
-
-sioverview::~sioverview()
-{
-    delete ui;
-}
