+++ sources/lusan/view/CMakeLists.txt
@@ -1,0 +1,2 @@
+include(${LUSAN}/view/common/CMakeLists.txt)
+include(${LUSAN}/view/si/CMakeLists.txt)
