+++ sources/lusan/view/common/ui/CMakeLists.txt
@@ -1,0 +1,3 @@
+list(APPEND LUSAN_UI
+    ${LUSAN}/view/common//ui/workspace.ui
+)
