+++ sources/lusan/view/si/CMakeLists.txt
@@ -1,0 +1,11 @@
+include("${LUSAN}/view/si/ui/CMakeLists.txt")
+
+list(APPEND LUSAN_SRC
+    ${LUSAN}/view/si/ServiceInterface.cpp
+    ${LUSAN}/view/si/SIOverview.cpp
+)
+
+list(APPEND LUSAN_HDR
+    ${LUSAN}/view/si/SIOverview.hpp
+    ${LUSAN}/view/si/SIOverview.hpp
+)
