+++ sources/lusan/view/si/ui/CMakeLists.txt
@@ -1,0 +1,3 @@
+list(APPEND LUSAN_UI
+    ${LUSAN}/view/si/ui/sioverview.ui
+)
